First commit
First commit. Fixed the default and authenticated user avatar images.

When I read Vite's docs during recording the video, it went over my head, and it was actually easy to import an image from the assets folder. I solved that in the layouts/MainLayout.vue and replaced the auth user's avatar URL in the Pinia stores/auth/index.ts.